% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{categorize_assessment_data}
\alias{categorize_assessment_data}
\title{Categorize assessment data into state/district/school levels}
\usage{
categorize_assessment_data(df, end_year)
}
\arguments{
\item{df}{Data frame from Excel file}

\item{end_year}{School year end}
}
\value{
List with state, district, school data frames
}
\description{
Categorize assessment data into state/district/school levels
}
\keyword{internal}
