% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{create_empty_result_list}
\alias{create_empty_result_list}
\title{Create empty result list}
\usage{
create_empty_result_list()
}
\value{
List with empty state, district, school data frames
}
\description{
Create empty result list
}
\keyword{internal}
