% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{download_assessment_file}
\alias{download_assessment_file}
\title{Download a single assessment file}
\usage{
download_assessment_file(url, end_year, level)
}
\arguments{
\item{url}{URL to download}

\item{end_year}{School year end}

\item{level}{Data level}
}
\value{
List with data frames
}
\description{
Download a single assessment file
}
\keyword{internal}
