% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{find_sheet}
\alias{find_sheet}
\title{Find a sheet by name pattern}
\usage{
find_sheet(sheets, patterns)
}
\arguments{
\item{sheets}{Vector of sheet names}

\item{patterns}{Vector of patterns to match (in order of preference)}
}
\value{
Matched sheet name or NULL
}
\description{
Find a sheet by name pattern
}
\keyword{internal}
