% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years for Michigan enrollment data}
\usage{
get_available_years()
}
\value{
A list with components:
\describe{
\item{min_year}{Earliest available year (1996)}
\item{max_year}{Most recent available year (2025)}
\item{description}{Human-readable description of the date range}
}
}
\description{
Returns the range of years for which enrollment data can be fetched
from the Michigan Center for Educational Performance and Information (CEPI).
}
\examples{
get_available_years()
}
