% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{read_assessment_file}
\alias{read_assessment_file}
\title{Read assessment file (Excel or ZIP)}
\usage{
read_assessment_file(filepath, is_zip, end_year)
}
\arguments{
\item{filepath}{Path to downloaded file}

\item{is_zip}{TRUE if ZIP file}

\item{end_year}{School year end}
}
\value{
List with data frames
}
\description{
Read assessment file (Excel or ZIP)
}
\keyword{internal}
