% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{read_headcount_sheet}
\alias{read_headcount_sheet}
\title{Read a headcount sheet from the Excel file}
\usage{
read_headcount_sheet(file_path, sheet_name, level, end_year = 2024)
}
\arguments{
\item{file_path}{Path to Excel file}

\item{sheet_name}{Name of sheet to read}

\item{level}{Data level (building, district, state)}

\item{end_year}{School year end (used to determine file format era)}
}
\value{
Data frame
}
\description{
Read a headcount sheet from the Excel file
}
\keyword{internal}
