% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{standardize_mi_grade}
\alias{standardize_mi_grade}
\title{Standardize Michigan grade levels}
\usage{
standardize_mi_grade(x)
}
\arguments{
\item{x}{Vector of grade values}
}
\value{
Standardized grade levels
}
\description{
Standardize Michigan grade levels
}
\keyword{internal}
