% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{standardize_mi_subgroup}
\alias{standardize_mi_subgroup}
\title{Standardize Michigan subgroup names}
\usage{
standardize_mi_subgroup(x)
}
\arguments{
\item{x}{Vector of subgroup names}
}
\value{
Standardized subgroup names
}
\description{
Standardize Michigan subgroup names
}
\keyword{internal}
