% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{standardize_mi_subject}
\alias{standardize_mi_subject}
\title{Standardize Michigan subject names}
\usage{
standardize_mi_subject(x)
}
\arguments{
\item{x}{Vector of subject names}
}
\value{
Standardized subject names
}
\description{
Standardize Michigan subject names
}
\keyword{internal}
